backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
psycopg[binary]
pydantic>=2.4.2
pydantic-settings>=2.0.3
bcrypt>=4.1.2
cffi>=1.16.0
python-jose[cryptography]==3.3.0
python-multipart>=0.0.6
reportlab>=4.0.7
Pillow>=10.2.0
python-dotenv>=1.0.1